Organize GUI and utility modules
