+++ gui/__init__.py
@@ -1,0 +1,1 @@
+from .gui import build_tabs
