--- themes.ini
@@ -1,66 +1,0 @@
-[DEFAULT]
-active_theme = light
-
-[light]
-bg = 255,255,255
-fg = 0,0,0
-panel_bg = 240,240,240
-button_bg = 230,230,230
-tab_bg = 230,230,230
-tab_active_bg = 210,210,210
-dropdown_bg = 230,230,230
-dropdown_fg = 0,0,0
-tree_bg = 255,255,255
-tree_fg = 0,0,0
-tree_select_bg = 200,200,200
-tree_select_fg = 0,0,0
-tree_heading_bg = 230,230,230
-tree_heading_fg = 0,0,0
-
-[dark]
-bg = 45,45,45
-fg = 220,220,220
-panel_bg = 55,55,55
-button_bg = 70,70,70
-tab_bg = 60,60,60
-tab_active_bg = 75,75,75
-dropdown_bg = 55,55,55
-dropdown_fg = 220,220,220
-tree_bg = 45,45,45
-tree_fg = 220,220,220
-tree_select_bg = 70,70,70
-tree_select_fg = 220,220,220
-tree_heading_bg = 60,60,60
-tree_heading_fg = 220,220,220
-
-[dark-blue]
-bg = 10,25,55
-fg = 173,216,230
-panel_bg = 20,35,70
-button_bg = 30,50,90
-tab_bg = 25,40,80
-tab_active_bg = 40,55,100
-dropdown_bg = 20,35,70
-dropdown_fg = 173,216,230
-tree_bg = 10,25,55
-tree_fg = 173,216,230
-tree_select_bg = 40,55,100
-tree_select_fg = 173,216,230
-tree_heading_bg = 25,40,80
-tree_heading_fg = 173,216,230
-
-[dark-green]
-bg = 10,55,25
-fg = 144,238,144
-panel_bg = 20,70,35
-button_bg = 30,90,50
-tab_bg = 25,80,45
-tab_active_bg = 40,100,60
-dropdown_bg = 20,70,35
-dropdown_fg = 144,238,144
-tree_bg = 10,55,25
-tree_fg = 144,238,144
-tree_select_bg = 40,100,60
-tree_select_fg = 144,238,144
-tree_heading_bg = 25,80,45
-tree_heading_fg = 144,238,144
