+++ utils/themes.ini
@@ -1,5 +1,5 @@
 [DEFAULT]
-active_theme = dark-green
+active_theme = light
 
 [light]
 bg = 255,255,255
@@ -64,4 +64,3 @@ tree_select_bg = 40,100,60
 tree_select_fg = 144,238,144
 tree_heading_bg = 25,80,45
 tree_heading_fg = 144,238,144
-
